Create onClick Pin handler, pass App state to child components
Place object is passed into Pin, and used in the onClick handler to
set the current place to currentPin. displayInformation is state
to handle the state of the footer information, and is changed to
True within the Pin onClick handler.

currentPin default state is a null dictionary to prevent the use in
Footer from erroring out